:recycle: Create hook to clean Main component
