+++ src/pages/Main/hook/index.ts
@@ -1,0 +1,1 @@
+export { default as useMain } from "./useMain";
